Add DNS provider cache
In the current implementation, DNS providers are called to list all
records on every loop. This is expensive in terms of number of requests
to the provider and may result in being rate limited.

With this, this PR introduces an optional generic caching mechanism at the provider
level, that re-uses the latest known list of records for a given time.

This prevents from expensive provider calls to list all records for each
object modification that does not change the actual record (annotations,
statuses, ingress routing, ...)

It introduces 2 trade-offs:

1. Any changes or corruption directly on the provider side will be
longer to detect and to resolve, up to the cache time

2. Any conflicting records in the DNS provider (such as a different
external-dns instance) injected during the cache validity will cause
the first iteration of the next reconcile loop to fail, and hence add a
delay until the next retry

Change-Id: I6e2424e68ab9bbb68e27712645249f83d169b422