Add the ability to create per-provider policies
Some providers may or may not support handling mixed record type (with
and without identifier). In the specific case of AWS Route 53, this is
not supported.

Hence, when 2 concurrent instances of external-dns runs with different
owners, if one has a record with identifier, and the other one without,
depending on the state of route 53, either records with identifier or
the ones without will be rejected to be created.

Having the ability to filter out those changes depending on the provider
capabilities will allow to fix this failure to register such domains and
block the whole integrations